{{cookiecutter.github_repository_name}}/requirements.txt
Django==1.8.4
Markdown==2.6.2
Pillow==2.9.0
argparse==1.4.0
beautifulsoup4==4.4.0
diff-match-patch==20121119
django-autoslug==1.8.0
django-debug-toolbar==1.3.2
django-extensions==1.5.7
django-filter==0.11.0
django-flat-theme==1.1.1
django-report-utils==0.3.13
django-versatileimagefield==1.0.4
djangorestframework==3.2.3
djangorestframework-filters==0.4.0
enum-compat==0.0.2
html5lib==0.9999999
jdcal==1.0
openpyxl==2.2.6
six==1.9.0
sqlparse==0.1.16
tablib==0.10.0
